refactor: split func call into separate file
